LS: Move all completion logic into `ide` module
commit-id:84062084
