+++ crates/cairo-lang-language-server/src/ide/completion/completions.rs
