+++ crates/cairo-lang-language-server/src/ide/mod.rs
@@ -1,2 +1,2 @@
-pub mod completions;
+pub mod completion;
 pub mod semantic_highlighting;
